+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 #Changelog
 ---
-####5.9.2_WIP
+####5.9.2
 - Added information upon sending a link of a GitHub repository branch
 - Added information upon sending a link of a GitHub issue or pull request
 - Added last updated time, wind direction and country to -w
 - Adapted imgur kicking to one-image-albums
 - Changed some GitHub messages
 - Fixed a rare error that occured in commit analyzing
+- Fixed error with imgur kicking
 
 ###5.9.1
 - Added info for reddit user links